requirements.txt
fastapi[all]
sqlalchemy
uvicorn
pydantic
alembic
psycopg2
# psycopg2-binary
python-multipart
python-jose[cryptography]
passlib
bcrypt